+++ src/aux-server/aux-web/aux-projector/interaction/ClickOperation/BuilderEmptyClickOperation.ts
@@ -25,8 +25,6 @@ export class BuilderEmptyClickOperation extends BaseEmptyClickOperation {
         appManager.simulationManager.primary.botPanel.restrictVisible(false);
 
         this.removeSelected();
-
-        this._game.gameView.$emit('onContextMenuHide');
     }
 
     async removeSelected() {